Do not change effective user ID and group ID in `Wserver`
The effect of the below code is to change the effective user and group ID of
the workers to match the user and group of the database file:
```ocaml
if Sys.unix then
 match bfile with
 | None -> ()
 | Some bfile ->
     let stat = Unix.stat bfile in
     let fuid = stat.Unix.st_uid in
     let fgid = stat.Unix.st_gid in
     let pgid = Unix.getgid () in
     let puid = Unix.getuid () in
     (* FIXME possible issue with effective uid/gid! *)
     (* see setuid, setgid man pages *)
     if puid <> fuid then Unix.setuid fuid;
     if pgid <> fgid then Unix.setgid fgid
```
In previous versions of GeneWeb, command-line options were used to specify
the user and group, instead of relying on the database file's permissions.
These options were removed around 2000.

I believe the intent of this code was to implement a form of privilege dropping.
On Unix systems, you often need special privileges to perform certain tasks, as
listening on ports below 1024. A common pattern is to start a server with root
privileges, bind your socket to a privileged port, and then drop those
privileges to a less-privileged user before handling client requests.

The current implementation is flawed for several reasons. This approach
is no longer recommended. `gwd` should not be run with effective root ID:
- I remove this code and prevent users from running `gwd` with too much privileges.
- I add documentation explaining how to open privileged port on Linux with root.